First Branch Commit
Adding comment for some properties that may or may not be necessary. They might be useful in the future but it depends on how we set up our code (ie. if we want to keep track of the winner based on the player or generally in the main code. Or if we wanted to do the same for which palyer's turn it is)